+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Add clarifying information to the `pcc secrets image-pull-secret` setup.
+
 - Modified the `aiohttp` minimum version to `3.11.12` and expanded `fastapi` to
   `>=0.115.6,<0.117.0` in order to align with `pipecat-ai`.